Move waitUntilTxPoolEmpty to global testing package
