+++ e2e/framework/helper.go
@@ -144,26 +144,6 @@ func WaitUntilTxPoolFilled(ctx context.Context, srv *TestServer, requiredNum uin
 	return res.(*txpoolProto.TxnPoolStatusResp), nil
 }
 
-// WaitUntilTxPoolEmpty waits until node has 0 transactions in txpool,
-// otherwise returns timeout
-func WaitUntilTxPoolEmpty(ctx context.Context, srv *TestServer) (*txpoolProto.TxnPoolStatusResp, error) {
-	clt := srv.TxnPoolOperator()
-	res, err := tests.RetryUntilTimeout(ctx, func() (interface{}, bool) {
-		subCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
-		defer cancel()
-		res, _ := clt.Status(subCtx, &empty.Empty{})
-		if res != nil && res.Length == 0 {
-			return res, false
-		}
-		return nil, true
-	})
-
-	if err != nil {
-		return nil, err
-	}
-	return res.(*txpoolProto.TxnPoolStatusResp), nil
-}
-
 // WaitUntilBlockMined waits until server mined block with bigger height than given height
 // otherwise returns timeout
 func WaitUntilBlockMined(ctx context.Context, srv *TestServer, desiredHeight uint64) (uint64, error) {
