+++ helper/tests/testing.go
@@ -4,12 +4,15 @@ import (
 	"context"
 	"crypto/ecdsa"
 	"errors"
+	"fmt"
+	txpoolOp "github.com/0xPolygon/polygon-sdk/txpool/proto"
 	"testing"
 	"time"
 
 	"github.com/0xPolygon/polygon-sdk/crypto"
 	"github.com/0xPolygon/polygon-sdk/types"
 	"github.com/stretchr/testify/assert"
+	empty "google.golang.org/protobuf/types/known/emptypb"
 )
 
 var (
@@ -49,4 +52,25 @@ func RetryUntilTimeout(ctx context.Context, f func() (interface{}, bool)) (inter
 	}()
 	res := <-resCh
 	return res.data, res.err
+}
+
+// WaitUntilTxPoolEmpty waits until node has 0 transactions in txpool,
+// otherwise returns timeout
+func WaitUntilTxPoolEmpty(ctx context.Context, client txpoolOp.TxnPoolOperatorClient) (*txpoolOp.TxnPoolStatusResp,
+	error) {
+	res, err := RetryUntilTimeout(ctx, func() (interface{}, bool) {
+		subCtx, cancel := context.WithTimeout(context.Background(), 5*time.Second)
+		defer cancel()
+		res, _ := client.Status(subCtx, &empty.Empty{})
+		if res != nil && res.Length == 0 {
+			return res, false
+		}
+		fmt.Printf("TxPool not empty, %d transactions remaining..\n", res.Length)
+		return nil, true
+	})
+
+	if err != nil {
+		return nil, err
+	}
+	return res.(*txpoolOp.TxnPoolStatusResp), nil
 }
